name colors and player popup changes
 - Designate temp/trial mods as a different color in both client and site
 - The player popup now pulls from the socket server if online first, then requests via HTTP if not found
 - Move some fns
 - Don't send raw password to server with login.php -- unset before it gets sent

+++ http_server/upload_level.php
@@ -75,7 +75,7 @@ try {
     }
 
     // sanity check: are they a guest?
-    $power = user_select_power($pdo, $user_id);
+    $power = (int) user_select_power($pdo, $user_id);
     if ($power <= 0) {
         $msg = 'Guests can\'t load or save levels. To access this feature, please create your own account.';
         throw new Exception($msg);
